Improve structure of icon-file types
- Make the allowed icon sizes format-dependent so that a different size
set can be configured for different formats
- Move the icon-file type configuration from the backend to the client,
because the icon-file type scheme is really just a guide/helper for user
input rather than a "business" rule to apply/enforce. (Benefits
justifying the cost involved in making it a "business" rule -- along
with related backend tests, etc. -- can be hardly seen at this point.)